3_requirements.txt
python-telegram-bot>=20.0
pandas>=1.3.0
openpyxl>=3.0.0
xlrd>=2.0.0
python-dotenv>=0.19.0
